[500320] Provides Tycho-based builds for coevolution repository
This change provides a work-in-progress build based on Maven/Tycho for
the co-evolution repository, but has a few issues, mainly related to
compiling the examples: both the APA and UML examples depend on various
parts of Capella and Kitalpha projects.

Given Capella builds have a dependency on the Diff/Merge project it is
required to ensure that the used Capella builds are based on the same
version of Diff/Merge core as coevolution is using internally.

As of now, this change does not build three example projects that depend
either directly or indirectly on capella.compare; the other example
projects have dependencies on Capella and Kitalpha, but no indirect
dependencies on Diff/Merge core.

Change-Id: Id49e644687d8dce1957b79dbd5dc98cae57b3313
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>